Add files via upload
Create a flow chart and algorithms
The goal of this program is to develop a flow chart for software that tracks student attendance in cohort meetings, 
records their scores and performance, and gathers their feedback on the program. 
Additionally, the application will analyze this data to predict if a student is likely to complete the program or drop out.